+++ Adj/Algebra.hs
@@ -1,3 +1,5 @@
 module Adj.Algebra (module Exports) where
 
 import Adj.Algebra.Semigroupoid as Exports
+import Adj.Algebra.Category as Exports
+import Adj.Algebra.Functor as Exports